enh: refactor package structure
